Refarctoring Search Script and added js dynamic switch input on type search
