+++ Site/Index.php
@@ -90,6 +90,7 @@ if(isset($_GET['MenuIndex']))
             require_once("../MedaLib/Function/SQL/SQLStatementExec.php");
             require_once("../MedaLib/Function/Generator/QuerySearchConstructor.php");
             require_once("../MedaLib/Function/Tool/RangeCheck.php");
+            require_once("Function/SearchQueryConstructor/CompanySearchQueryConstr.php");
             break;
 
         //Country
@@ -98,6 +99,7 @@ if(isset($_GET['MenuIndex']))
             require_once("Function/AccessLevelCheck.php");
             require_once("../MedaLib/Function/Generator/QuerySearchConstructor.php");
             require_once("../MedaLib/Function/Tool/RangeCheck.php");
+            require_once("Function/SearchQueryConstructor/CountrySearchQueryConstr.php");
             break;
 
         //Employee
@@ -106,6 +108,7 @@ if(isset($_GET['MenuIndex']))
             require_once("Function/AccessLevelCheck.php");
             require_once("../MedaLib/Function/Generator/QuerySearchConstructor.php");
             require_once("../MedaLib/Function/Tool/RangeCheck.php");
+            require_once("Function/SearchQueryConstructor/EmployeeSearchQueryConstr.php");
             break;
 
         //Employee Position
@@ -114,6 +117,7 @@ if(isset($_GET['MenuIndex']))
             require_once("Function/AccessLevelCheck.php");
             require_once("../MedaLib/Function/Generator/QuerySearchConstructor.php");
             require_once("../MedaLib/Function/Tool/RangeCheck.php");
+            require_once("Function/SearchQueryConstructor/EmployeePositionSearchQueryConstr.php");
             break;
 
         //Job
@@ -122,6 +126,7 @@ if(isset($_GET['MenuIndex']))
             require_once("Function/AccessLevelCheck.php");
             require_once("../MedaLib/Function/Generator/QuerySearchConstructor.php");
             require_once("../MedaLib/Function/Tool/RangeCheck.php");
+            require_once("Function/SearchQueryConstructor/JobSearchQueryConstr.php");
             break;
 
         //Shareholder
@@ -130,6 +135,7 @@ if(isset($_GET['MenuIndex']))
             require_once("Function/AccessLevelCheck.php");
             require_once("../MedaLib/Function/Generator/QuerySearchConstructor.php");
             require_once("../MedaLib/Function/Tool/RangeCheck.php");
+            require_once("Function/SearchQueryConstructor/ShareholderSearchQueryConstr.php");
             break;
 
         //Customer
@@ -138,6 +144,7 @@ if(isset($_GET['MenuIndex']))
             require_once("Function/AccessLevelCheck.php");
             require_once("../MedaLib/Function/Generator/QuerySearchConstructor.php");
             require_once("../MedaLib/Function/Tool/RangeCheck.php");
+            require_once("Function/SearchQueryConstructor/CustomerSearchQueryConstr.php");
             break;
 
         //County
@@ -146,6 +153,7 @@ if(isset($_GET['MenuIndex']))
             require_once("Function/AccessLevelCheck.php");
             require_once("../MedaLib/Function/Generator/QuerySearchConstructor.php");
             require_once("../MedaLib/Function/Tool/RangeCheck.php");
+            require_once("Function/SearchQueryConstructor/CountySearchQueryConstr.php");
             break;
 
         default:
