+++ Site/Output/Retriever/CountryRetriever.php
@@ -1,23 +1,4 @@
 <?php
-function CountrySearchConstructor(string &$InsSearchTypeQuery, string &$IniSearchType) : void
-{
-	switch($IniSearchType)
-	{
-		case $GLOBALS['COUNTRY_SEARCH_TYPE']['COUNTRY_TITLE']["NAME"]:
-		{
-			$InsSearchTypeQuery = "COUN_DATA_TITLE";
-			break;
-		}
-
-		default:
-		{
-			$InsSearchTypeQuery = "COUN_DATA_TITLE";
-			break;
-		}
-	}
-}
-
-
 function CountryRetriever(ME_CDBConnManager &$InrConn, ME_CLogHandle &$InrLogHandle, int $IniUserAccess, int $IniAvail)
 {
 	if(CheckAccessRange($IniUserAccess) &&
