+++ Site/Output/Retriever/JobRetriever.php
@@ -1,58 +1,4 @@
 <?php
-function JobSearchConstructor(string &$InsSearchTypeQuery, string &$IniSearchType) : void
-{
-	switch($IniSearchType)
-	{
-		case $GLOBALS['JOB_SEARCH_TYPE']['JOB_TITLE']["NAME"]:
-		{
-			$InsSearchTypeQuery = "JOB_DATA_Title";
-			break;
-		}
-
-		case $GLOBALS['JOB_SEARCH_TYPE']['JOB_DATE']["NAME"]:
-		{
-			$InsSearchTypeQuery = "JOB_DATA_Date";
-			break;
-		}
-
-		case $GLOBALS['JOB_SEARCH_TYPE']['JOB_PRICE']["NAME"]:
-		{
-			$InsSearchTypeQuery = "JOB_DATA_Price";
-			break;
-		}
-
-		case $GLOBALS['JOB_SEARCH_TYPE']['JOB_PIA']["NAME"]:
-		{
-			$InsSearchTypeQuery = "JOB_DATA_PIA";
-			break;
-		}
-
-		case $GLOBALS['JOB_SEARCH_TYPE']['JOB_EXPENSES']["NAME"]:
-		{
-			$InsSearchTypeQuery = "JOB_DATA_Expenses";
-			break;
-		}
-
-		case $GLOBALS['JOB_SEARCH_TYPE']['JOB_DAMAGE']["NAME"]:
-		{
-			$InsSearchTypeQuery = "JOB_DATA_Damage";
-			break;
-		}
-
-		case $GLOBALS['JOB_SEARCH_TYPE']['JOB_COMPANY_TITLE']["NAME"]:
-		{
-			$InsSearchTypeQuery = "COMPANY_DATA_Title";
-			break;
-		}
-
-		default:
-		{
-			$InsSearchTypeQuery = "JOB_DATA_Title";
-			break;
-		}
-	}
-}
-
 function JobRetriever(ME_CDBConnManager &$InrConn, ME_CLogHandle &$InrLogHandle, int $IniUserAccess, int $IniAvail)
 {
 	if(CheckAccessRange($IniUserAccess) &&
