+++ Site/Struct/Module/Overview/HTMLModule/HTMLModuleCountryOverview.php
@@ -26,7 +26,7 @@ function HTMLCountryOverviewDataBlock(mysqli_result &$InrResult, ME_CLogHandle &
     $sHTMLGeneratedSelectStructure = "";
     $sSearchTypeSelected = isset($_GET[$sSearchSelectStructName]) ? $_GET[$sSearchSelectStructName] : "";
 
-    HTMLGenerateSelectStructure($sHTMLGeneratedSelectStructure, $sSearchSelectStructName, $GLOBALS['COUNTRY_SEARCH_TYPE'], $sSearchTypeSelected, "QueryDataType", "onchange", "CountryQueryDataType()");
+    HTMLGenerateSelectStructure($sHTMLGeneratedSelectStructure, $sSearchSelectStructName, $GLOBALS['COUNTRY_SEARCH_TYPE'], $sSearchTypeSelected, "QueryDataType", "onchange", "CounQueryDataType()");
 
     //The toolbar for the buttons (tools)
     printf("
