+++ Site/Struct/Module/Overview/HTMLModule/HTMLModuleShareholderOverview.php
@@ -26,7 +26,7 @@ function HTMLShareholderDataBlock(mysqli_result &$InrResult, ME_CLogHandle &$Inr
     $sHTMLGeneratedSelectStructure = "";
     $sSearchTypeSelected = isset($_GET[$sSearchSelectStructName]) ? $_GET[$sSearchSelectStructName] : "";
 
-    HTMLGenerateSelectStructure($sHTMLGeneratedSelectStructure, $sSearchSelectStructName, $GLOBALS['EMPLOYEE_POSITION_SEARCH_TYPE'], $sSearchTypeSelected);
+    HTMLGenerateSelectStructure($sHTMLGeneratedSelectStructure, $sSearchSelectStructName, $GLOBALS['SHAREHOLDER_SEARCH_TYPE'], $sSearchTypeSelected, "QueryDataType", "onchange", "ShareQueryDataType()");
 
 	//The toolbar for the buttons (tools)
 	printf("
@@ -36,7 +36,7 @@ function HTMLShareholderDataBlock(mysqli_result &$InrResult, ME_CLogHandle &$Inr
 		</a>
 		<form action='.' method='get'>
 			<input type='hidden' name='MenuIndex' value='%d'><label>Search by%s</label>
-			<label>Query</label><input type='text' name='SearchQuery' value='%s'>
+			<label>Query</label><input type='text' id='QueryInput' name='SearchQuery' value='%s'>
 			<button>submit</button>
 		</form>
 	</div>",
